Finalize pt2

+++ src/main/java/org/restaurant/models/Ingredient.java
@@ -1,15 +1,19 @@
 package org.restaurant.models;
 
+import org.restaurant.DAO.IngredientHistoryDAO;
+import org.restaurant.DAO.StockDAO;
+
 import java.time.LocalDateTime;
+import java.util.List;
 
 public class Ingredient {
     private int id_ingredient;
     private String name;
-    private int unit_price;
-    private Unit unit ;
+    private double unit_price;
+    private Unit unit;
     private LocalDateTime update_dateTime;
 
-    public Ingredient(int id_ingredient, String name, int unit_price, Unit unit, LocalDateTime update_dateTime) {
+    public Ingredient(int id_ingredient, String name, double unit_price, Unit unit, LocalDateTime update_dateTime) {
         this.id_ingredient = id_ingredient;
         this.name = name;
         this.unit_price = unit_price;
@@ -37,11 +41,11 @@ public class Ingredient {
         this.name = name;
     }
 
-    public int getUnit_price() {
+    public double getUnit_price() {
         return unit_price;
     }
 
-    public void setUnit_price(int unit_price) {
+    public void setUnit_price(double unit_price) {
         this.unit_price = unit_price;
     }
 
@@ -60,4 +64,10 @@ public class Ingredient {
     public void setUpdate_dateTime(LocalDateTime update_dateTime) {
         this.update_dateTime = update_dateTime;
     }
+
+
+
+
+
+
 }
